requirements.txt
langchain
langchain-openai
langchain-chroma
openai
tiktoken
chromadb
unstructured
pypdf
jq
fastapi
uvicorn
pydantic
python-dotenv 